Add packetlog-test to test table performance
Include a test to check how many packets we can process in a timeframe
of 10 seconds. The test first puts the requested amount of records into
the table, and then keeps updating a final entry in the table. The idea
is to know how many packets it can process in 10 seconds for a given
table size in case it has to search through the entire table to find
the matching record to be updated.